Update SCU3_HD_patch.py
Removed file path construction, as we're already looking in the cwd based on requirement to place script in executable folder. If-statement handles check.

This allows the script to work in Linux for Steam versions of the game.

This change does not affect its use in Windows, as Python OS libraries are agnostic.